move distribution_learners files under qgan folder
