+++ qiskit_machine_learning/algorithms/distribution_learners/qgan/__init__.py
@@ -10,14 +10,14 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-""" Distribution Learners Package """
+""" QGAN Package """
 
-from .qgan import (DiscriminativeNetwork,
-                   GenerativeNetwork,
-                   NumPyDiscriminator,
-                   PyTorchDiscriminator,
-                   QuantumGenerator,
-                   QGAN)
+from .discriminative_network import DiscriminativeNetwork
+from .generative_network import GenerativeNetwork
+from .numpy_discriminator import NumPyDiscriminator
+from .pytorch_discriminator import PyTorchDiscriminator
+from .quantum_generator import QuantumGenerator
+from .qgan import QGAN
 
 __all__ = [
     'DiscriminativeNetwork',
