+++ qiskit_machine_learning/algorithms/distribution_learners/qgan/_pytorch_discriminator_net.py
