+++ qiskit_machine_learning/algorithms/distribution_learners/qgan/discriminative_network.py
