+++ qiskit_machine_learning/algorithms/distribution_learners/qgan/generative_network.py
