+++ qiskit_machine_learning/algorithms/distribution_learners/qgan/numpy_discriminator.py
