+++ qiskit_machine_learning/algorithms/distribution_learners/qgan/pytorch_discriminator.py
