+++ test/algorithms/distribution_learners/qgan/test_qgan.py
